Large commit today...
* Child pid's State is tracked.
* start-civ-server closes down and terminates if civserver dies.
* Port is released if civserver child dies.
* No longer use File::Tail::App, instead using File::Tail
* Log file read at 5 second intervals

All we need now is to dump the collected data and then the web front end can be written.

I suspect we'll have to handle more server messages in _parse_log, especially to deal with ai.

we need to be able to resume a halted server by loading up the last dump and starting the server with the latest savegame.

Really getting there now, everything seems very stable.